+++ src/papilo/core/postsolve/Postsolve.hpp
@@ -475,14 +475,14 @@ Postsolve<REAL>::copy_from_reduced_to_original(
           postsolveStorage.presolveOptions.calculate_basis_for_dual;
       int reduced_rows = (int)reducedSolution.dual.size();
 
-      assert( (int)  reducedSolution.dual.size() == reduced_rows );
+      assert( reducedSolution.dual.size() == reduced_rows );
       originalSolution.dual.clear();
       originalSolution.dual.resize( postsolveStorage.nRowsOriginal );
       for( int k = 0; k < reduced_rows; k++ )
          originalSolution.dual[postsolveStorage.origrow_mapping[k]] =
              reducedSolution.dual[k];
 
-      assert( (int) reducedSolution.reducedCosts.size() == reduced_columns );
+      assert( reducedSolution.reducedCosts.size() == reduced_columns );
       originalSolution.reducedCosts.clear();
       originalSolution.reducedCosts.resize( postsolveStorage.nColsOriginal );
       for( int k = 0; k < reduced_columns; k++ )
@@ -497,7 +497,7 @@ Postsolve<REAL>::copy_from_reduced_to_original(
          originalSolution.varBasisStatus[postsolveStorage.origcol_mapping[k]] =
              reducedSolution.varBasisStatus[k];
 
-      assert( (int) reducedSolution.rowBasisStatus.size() == reduced_rows );
+      assert( reducedSolution.rowBasisStatus.size() == reduced_rows );
 
       originalSolution.rowBasisStatus.clear();
       originalSolution.rowBasisStatus.resize( postsolveStorage.nRowsOriginal,
